Add FirewalledRule
I added a FirewalledRule which can be extended in order to restrict
relaying of meta-transactions to organization workers.
The rule that is to be restricted must extend the FirewalledRule and add
the modifier `firewalled` to all relevant functions.

For testing I created two test contract files. One that uses the
modifier, and one that calls the one that uses the modifier. Required in
order to simulate rules calling on each other and checking for the
relayer.

Added a step to manually install npx inside docker until the docker
image is fixed. The problem is reported and confirmed at
https://github.com/crytic/eth-security-toolbox/issues/8

Fixes #199